// lib/types.ts
export type Exercise = {
  id: number;
  name_en: string;
  name_fa: string;
  description_fa: string;
  muscle_group_fa: string;
  difficulty_fa: string;
  equipment_fa: string;
  image_url: string;
  video_url: string;
  instructions_fa: string[];
  dos_fa: string[];
  donts_fa: string[];
  breathing_fa: string;
  safety_fa: string;
};